Shader abstraction

+++ Cabrium/src/Cabrium/Platform/Windows/WindowsUtils.cpp
@@ -6,6 +6,6 @@
 
 namespace cabrium {
 
-float Time::getTime() { return glfwGetTime(); }
+float Time::getTime() { return (float) glfwGetTime(); }
 
 } // namespace cabrium
